simplemfa/requirements.txt
django-twilio>=0.13.0

